move fileUserDataProvider to platform and use it in shared process
