+++ src/vs/code/electron-browser/sharedProcess/sharedProcessMain.ts
@@ -97,6 +97,7 @@ import { IUriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentity'
 import { UriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentityService';
 import { IPCFileSystemProvider } from 'vs/platform/files/common/ipcFileSystemProvider';
 import { isLinux } from 'vs/base/common/platform';
+import { FileUserDataProvider } from 'vs/platform/userData/common/fileUserDataProvider';
 
 class SharedProcessMain extends Disposable {
 
@@ -206,7 +207,8 @@ class SharedProcessMain extends Disposable {
 		const diskFileSystemProvider = this._register(new DiskFileSystemProvider(logService));
 		fileService.registerProvider(Schemas.file, diskFileSystemProvider);
 
-		const userDataFileSystemProvider = this._register(new IPCFileSystemProvider(mainProcessService.getChannel('localFilesystem'), { pathCaseSensitive: isLinux }));
+		const localFileSystemProviderFromMain = this._register(new IPCFileSystemProvider(mainProcessService.getChannel('localFilesystem'), { pathCaseSensitive: isLinux }));
+		const userDataFileSystemProvider = this._register(new FileUserDataProvider(Schemas.file, localFileSystemProviderFromMain, Schemas.userData, logService));
 		fileService.registerProvider(Schemas.userData, userDataFileSystemProvider);
 
 		// Configuration
